完结

+++ src/main/java/com/rh/i_mes_plus/dto/UmsWmsAreaDTO.java
@@ -11,7 +11,7 @@ public class UmsWmsAreaDTO {
     private String arFatherSn;
     private String connector;
     private String arState;
-    private Long arArea;
+    private Integer arArea;
     private Long arAreaUnit;
     private Long arCubage;
     private Long arCubageUnit;
